requirements.txt
ipywidgets==7.6.3
matplotlib==3.4.1
numpy==1.20.2
opencv-python==4.5.1.48
pandas==1.2.4
Pillow==8.2.0
pytesseract==0.3.7
scipy==1.6.2
tensorboard==2.4.1
tqdm==4.60.0

torch==1.8.1+cu111
torchvision==0.9.1+cu111
